monitor: avoid signed-shift UB in the s16 widening path
Left-shifting a negative signed value is undefined before C++20 and the
monitor builds as C++17. widen_s16_to_s32() and convert_to_pipe_format()
now shift through the unsigned type, which is value-preserving on any
two's-complement target and compiles to the same instruction. The unit
tests computed their expected values the same way, so those are updated
too, and convert_to_pipe_format() no longer hands memcpy a null source
for a zero-length copy.

Also wires the control-thread reaping regression test into the native
test runner, where it had never been invoked, and corrects three comments
that still described the disarm fade with its former duration.

+++ core/monitor/autostream_monitor_utils.h
@@ -62,7 +62,12 @@ std::array<std::uint8_t, 44> build_wav_header(int rate,
 // without an ALSA device.
 inline constexpr std::int32_t widen_s16_to_s32(std::int16_t sample)
 {
-    return static_cast<std::int32_t>(sample) << 16;
+    // Shift through the unsigned type: left-shifting a negative signed value
+    // is undefined before C++20, and this is built as C++17. The round trip
+    // is value-preserving on any two's-complement target and compiles to the
+    // same instruction.
+    return static_cast<std::int32_t>(
+        static_cast<std::uint32_t>(static_cast<std::int32_t>(sample)) << 16);
 }
